Move MavenInstaller to models
