+++ modules/mining-pipeline/github-project-miner/build.gradle
@@ -1,10 +1,9 @@
 dependencies {
     compile project(":mining-pipeline")
     compile project(":models:java-project")
+    compile project(":models:maven-installer")
 
-    compile group: "org.apache.maven.shared", name: "maven-invoker", version: mavenInvokerVersion
     compile group: "org.kohsuke", name: "github-api", version: kohsukeVersion
-    compile group: "org.zeroturnaround", name: "zt-zip", version: ztZipVersion
 
     testCompile group: "com.nhaarman", name: "mockito-kotlin", version: mockitoKotlinVersion
 }
