+++ modules/mining-pipeline/java-maven-project-compiler/build.gradle
@@ -5,4 +5,6 @@ dependencies {
 
     compile group: "org.apache.maven.shared", name: "maven-invoker", version: mavenInvokerVersion
     compile group: "org.zeroturnaround", name: "zt-zip", version: ztZipVersion
+
+    testCompile project(":models:maven-installer")
 }
