+++ modules/models/maven-installer/gradle.properties
@@ -1,0 +1,2 @@
+mavenInvokerVersion  = 3.0.0
+ztZipVersion         = 1.13
